AisEncoder: Add configurable number of lead-in symbols
At least in my setup here with the B200, transmission needs
to start well ahead of any actual useful information, so we
add the option of transmitting 'lead-in' symbols.

A proper solution would require to understand the issue in more
detail.   rather than transmitting valid 0 symbols at full power,
we should have the transmitter start up (at zero/low power) and
then ramp up until the first actual symbol of the burst is to be
transmitted.